microservices/kafka-service/consumer/requirements.txt
kafka-python==2.0.2
flake8==6.0.0
python-dotenv==0.19.0
requests